requirements.txt
streamlit 
pdf2image 
pytesseract 
python-docx 
pillow
opencv-python
numpy
pdfplumber
googletrans==4.0.0-rc1
deep-translator
pdfplumber 
reportlab
PyPDF2

